Making rubocop happy
